[BUG] Fix program segfaulting if no parameters were given and fix the program misbehaving if too many arguments were given.

+++ main.cc
@@ -1,16 +1,24 @@
 #include "libFK.hh"
 #include "src.hh"
-#include <iostream>
 #include <filesystem>
 #include <string>
 
 namespace fs = std::filesystem;
 
 int main(int argc, char *argv[]) {
-    fs::create_directory("out");
-    src::outputNeededLibraries(&*argv[1], "out/neededLibraries.txt");
-    src::generateInstallScript("out/libNotFound.txt");
-    src::handleBuildingAndRunningTheProgram("out/runLibraryInstallScripts.sh", "out/buildFlags.txt", argc, argv[1]);
-    system("rm -rf out");
-    return system("rm -rf a.out");;
+    if (argc < 2) {
+        fk::msg(1, "No input file specified.");
+    } else if (argc > 2) {
+        fk::msg(1, "Too many inputs specified.");
+    }
+    else {
+        fs::create_directory("out");
+        src::outputNeededLibraries(&*argv[1], "out/neededLibraries.txt");
+        src::generateInstallScript("out/libNotFound.txt");
+        src::handleBuildingAndRunningTheProgram("out/runLibraryInstallScripts.sh", "out/buildFlags.txt", argc, argv[1]);
+        system("rm -rf out");
+        system("rm -rf a.out");
+    }
+
+    return 0;
 }
